Deleting a trainer or removing the trainer role now works as expected. It shows an error if he has lessons assigned to him.

+++ src/Ilc.Web/Scripts/app/grid/Roles.js
@@ -61,11 +61,15 @@
                         icon: 'images/web/remove.png',
                         tooltip: Ilc.resources.Manager.getResourceString('common.delete'),
                         handler: function (grid, rowIndex, colIndex, item, e, record) {
+                            if (me.userRoles) {
+                                me.fireEvent('delete', me, record);
+                            } else {
+                                record.destroy({
+                                    success: me.onDeleteSuccess,
+                                    failure: me.onDeleteFailure
+                                });
+                            }
                             
-                            record.destroy({
-                                success: me.onDeleteSuccess,
-                                failure: me.onDeleteFailure
-                            });
                             
                         }
                     }
